reactor: react to modules
abstract database operation to independent module and make it interface
